requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cloudinary==1.29.0
Django==4.0.5
django-location-field==2.1.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
phonenumbers==8.12.51
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
